+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { IntegrationBreadcrumb } from './integration_breadcrumb';
+export { CreatePackagePolicyPage } from './single_page_layout';
