+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/components/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { CreatePackagePolicyPageLayout } from './layout';
+export { PackagePolicyInputPanel } from './package_policy_input_panel';
+export { PackagePolicyInputVarField } from './package_policy_input_var_field';
+export { PostInstallAddAgentModal } from './post_install_add_agent_modal';
