+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/components/multi_text_input.test.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import { fireEvent, act } from '@testing-library/react';
 
-import { createFleetTestRendererMock } from '../../../../../../mock';
+import { createFleetTestRendererMock } from '../../../../../../../mock';
 
 import { MultiTextInput } from './multi_text_input';
