+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/components/package_policy_input_config.tsx
@@ -6,24 +6,18 @@
  */
 
 import React, { useState, Fragment, memo, useMemo } from 'react';
-import ReactMarkdown from 'react-markdown';
 import styled from 'styled-components';
 import { FormattedMessage } from '@kbn/i18n-react';
 import {
   EuiFlexGrid,
   EuiFlexGroup,
   EuiFlexItem,
-  EuiSwitch,
   EuiText,
   EuiSpacer,
   EuiButtonEmpty,
 } from '@elastic/eui';
 
-import type {
-  NewPackagePolicyInputStream,
-  RegistryStream,
-  RegistryVarsEntry,
-} from '../../../../types';
+import type { NewPackagePolicyInput, RegistryVarsEntry } from '../../../../../types';
 import type { PackagePolicyConfigValidationResults } from '../services';
 import { isAdvancedVar, validationHasErrors } from '../services';
 
@@ -33,32 +27,34 @@ const FlexItemWithMaxWidth = styled(EuiFlexItem)`
   max-width: calc(50% - ${(props) => props.theme.eui.euiSizeL});
 `;
 
-export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
-  packageInputStream: RegistryStream;
-  packagePolicyInputStream: NewPackagePolicyInputStream;
-  updatePackagePolicyInputStream: (updatedStream: Partial<NewPackagePolicyInputStream>) => void;
-  inputStreamValidationResults: PackagePolicyConfigValidationResults;
+export const PackagePolicyInputConfig: React.FunctionComponent<{
+  hasInputStreams: boolean;
+  packageInputVars?: RegistryVarsEntry[];
+  packagePolicyInput: NewPackagePolicyInput;
+  updatePackagePolicyInput: (updatedInput: Partial<NewPackagePolicyInput>) => void;
+  inputVarsValidationResults: PackagePolicyConfigValidationResults;
   forceShowErrors?: boolean;
 }> = memo(
   ({
-    packageInputStream,
-    packagePolicyInputStream,
-    updatePackagePolicyInputStream,
-    inputStreamValidationResults,
+    hasInputStreams,
+    packageInputVars,
+    packagePolicyInput,
+    updatePackagePolicyInput,
+    inputVarsValidationResults,
     forceShowErrors,
   }) => {
     // Showing advanced options toggle state
-    const [isShowingAdvanced, setIsShowingAdvanced] = useState<boolean>();
+    const [isShowingAdvanced, setIsShowingAdvanced] = useState<boolean>(false);
 
     // Errors state
-    const hasErrors = forceShowErrors && validationHasErrors(inputStreamValidationResults);
+    const hasErrors = forceShowErrors && validationHasErrors(inputVarsValidationResults);
 
     const requiredVars: RegistryVarsEntry[] = [];
     // eslint-disable-next-line react-hooks/exhaustive-deps
     const advancedVars: RegistryVarsEntry[] = [];
 
-    if (packageInputStream.vars && packageInputStream.vars.length) {
-      packageInputStream.vars.forEach((varDef) => {
+    if (packageInputVars) {
+      packageInputVars.forEach((varDef) => {
         if (isAdvancedVar(varDef)) {
           advancedVars.push(varDef);
         } else {
@@ -70,9 +66,9 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
     const advancedVarsWithErrorsCount: number = useMemo(
       () =>
         advancedVars.filter(
-          ({ name: varName }) => inputStreamValidationResults?.vars?.[varName]?.length
+          ({ name: varName }) => inputVarsValidationResults.vars?.[varName]?.length
         ).length,
-      [advancedVars, inputStreamValidationResults?.vars]
+      [advancedVars, inputVarsValidationResults.vars]
     );
 
     return (
@@ -81,24 +77,26 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
           <EuiFlexGroup gutterSize="none" alignItems="flexStart">
             <EuiFlexItem grow={1} />
             <EuiFlexItem grow={5}>
-              <EuiSwitch
-                label={packageInputStream.title}
-                disabled={packagePolicyInputStream.keep_enabled}
-                checked={packagePolicyInputStream.enabled}
-                onChange={(e) => {
-                  const enabled = e.target.checked;
-                  updatePackagePolicyInputStream({
-                    enabled,
-                  });
-                }}
-              />
-              {packageInputStream.description ? (
-                <Fragment>
+              <EuiText>
+                <h4>
+                  <FormattedMessage
+                    id="xpack.fleet.createPackagePolicy.stepConfigure.inputSettingsTitle"
+                    defaultMessage="Settings"
+                  />
+                </h4>
+              </EuiText>
+              {hasInputStreams ? (
+                <>
                   <EuiSpacer size="s" />
-                  <EuiText size="s" color="subdued">
-                    <ReactMarkdown source={packageInputStream.description} />
+                  <EuiText color="subdued" size="s">
+                    <p>
+                      <FormattedMessage
+                        id="xpack.fleet.createPackagePolicy.stepConfigure.inputSettingsDescription"
+                        defaultMessage="The following settings are applicable to all inputs below."
+                      />
+                    </p>
                   </EuiText>
-                </Fragment>
+                </>
               ) : null}
             </EuiFlexItem>
           </EuiFlexGroup>
@@ -106,12 +104,9 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
         <FlexItemWithMaxWidth>
           <EuiFlexGroup direction="column" gutterSize="m">
             {requiredVars.map((varDef) => {
-              if (!packagePolicyInputStream?.vars) return null;
               const { name: varName, type: varType } = varDef;
-              const varConfigEntry = packagePolicyInputStream.vars?.[varName];
-              const value = varConfigEntry?.value;
-              const frozen = varConfigEntry?.frozen ?? false;
-
+              if (!packagePolicyInput.vars) return;
+              const { value, frozen } = packagePolicyInput.vars[varName];
               return (
                 <EuiFlexItem key={varName}>
                   <PackagePolicyInputVarField
@@ -119,9 +114,9 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
                     value={value}
                     frozen={frozen}
                     onChange={(newValue: any) => {
-                      updatePackagePolicyInputStream({
+                      updatePackagePolicyInput({
                         vars: {
-                          ...packagePolicyInputStream.vars,
+                          ...packagePolicyInput.vars,
                           [varName]: {
                             type: varType,
                             value: newValue,
@@ -129,7 +124,7 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
                         },
                       });
                     }}
-                    errors={inputStreamValidationResults?.vars![varName]}
+                    errors={inputVarsValidationResults.vars![varName]}
                     forceShowErrors={forceShowErrors}
                   />
                 </EuiFlexItem>
@@ -138,6 +133,7 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
             {advancedVars.length ? (
               <Fragment>
                 <EuiFlexItem>
+                  {/* Wrapper div to prevent button from going full width */}
                   <EuiFlexGroup justifyContent="spaceBetween" alignItems="center">
                     <EuiFlexItem grow={false}>
                       <EuiButtonEmpty
@@ -167,19 +163,17 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
                 </EuiFlexItem>
                 {isShowingAdvanced
                   ? advancedVars.map((varDef) => {
-                      if (!packagePolicyInputStream.vars) return null;
                       const { name: varName, type: varType } = varDef;
-                      const value = packagePolicyInputStream.vars?.[varName]?.value;
-
+                      const value = packagePolicyInput.vars?.[varName]?.value;
                       return (
                         <EuiFlexItem key={varName}>
                           <PackagePolicyInputVarField
                             varDef={varDef}
                             value={value}
                             onChange={(newValue: any) => {
-                              updatePackagePolicyInputStream({
+                              updatePackagePolicyInput({
                                 vars: {
-                                  ...packagePolicyInputStream.vars,
+                                  ...packagePolicyInput.vars,
                                   [varName]: {
                                     type: varType,
                                     value: newValue,
@@ -187,7 +181,7 @@ export const PackagePolicyInputStreamConfig: React.FunctionComponent<{
                                 },
                               });
                             }}
-                            errors={inputStreamValidationResults?.vars![varName]}
+                            errors={inputVarsValidationResults.vars![varName]}
                             forceShowErrors={forceShowErrors}
                           />
                         </EuiFlexItem>
