+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/components/package_policy_input_panel.tsx
@@ -24,7 +24,7 @@ import type {
   PackagePolicyInputStream,
   RegistryInput,
   RegistryStream,
-} from '../../../../types';
+} from '../../../../../types';
 import type { PackagePolicyInputValidationResults } from '../services';
 import { hasInvalidButRequiredVar, countValidationErrors } from '../services';
