+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/components/package_policy_input_var_field.tsx
@@ -21,7 +21,7 @@ import styled from 'styled-components';
 
 import { CodeEditor } from '@kbn/kibana-react-plugin/public';
 
-import type { RegistryVarsEntry } from '../../../../types';
+import type { RegistryVarsEntry } from '../../../../../types';
 
 import { MultiTextInput } from './multi_text_input';
