+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/services/has_invalid_but_required_var.ts
@@ -7,7 +7,7 @@
 
 import { safeLoad } from 'js-yaml';
 
-import type { PackagePolicyConfigRecord, RegistryVarsEntry } from '../../../../types';
+import type { PackagePolicyConfigRecord, RegistryVarsEntry } from '../../../../../types';
 
 import { validatePackagePolicyConfig } from '.';
