+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/services/index.ts
@@ -11,10 +11,10 @@ export type {
   PackagePolicyValidationResults,
   PackagePolicyConfigValidationResults,
   PackagePolicyInputValidationResults,
-} from '../../../../services';
+} from '../../../../../services';
 export {
   validatePackagePolicy,
   validatePackagePolicyConfig,
   validationHasErrors,
   countValidationErrors,
-} from '../../../../services';
+} from '../../../../../services';
