+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/services/is_advanced_var.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { RegistryVarsEntry } from '../../../../types';
+import type { RegistryVarsEntry } from '../../../../../types';
 
 export const isAdvancedVar = (varDef: RegistryVarsEntry): boolean => {
   if (varDef.show_user || (varDef.required && varDef.default === undefined)) {
