+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/step_configure_package.tsx
@@ -15,9 +15,9 @@ import {
 } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
 
-import type { PackageInfo, NewPackagePolicy, NewPackagePolicyInput } from '../../../types';
-import { Loading } from '../../../components';
-import { getStreamsForInputType, doesPackageHaveIntegrations } from '../../../services';
+import type { PackageInfo, NewPackagePolicy, NewPackagePolicyInput } from '../../../../types';
+import { Loading } from '../../../../components';
+import { getStreamsForInputType, doesPackageHaveIntegrations } from '../../../../services';
 
 import type { PackagePolicyValidationResults } from './services';
 import { PackagePolicyInputPanel } from './components';
