+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/step_select_hosts.tsx
@@ -10,12 +10,12 @@ import type { EuiTabbedContentTab } from '@elastic/eui';
 import { EuiTabbedContent } from '@elastic/eui';
 import styled from 'styled-components';
 
-import { useGetAgentPolicies } from '../../../hooks';
-import type { AgentPolicy, NewAgentPolicy, PackageInfo } from '../../../types';
-import { AgentPolicyIntegrationForm } from '../components';
-import type { ValidationResults } from '../components/agent_policy_validation';
+import { useGetAgentPolicies } from '../../../../hooks';
+import type { AgentPolicy, NewAgentPolicy, PackageInfo } from '../../../../types';
+import { AgentPolicyIntegrationForm } from '../../components';
+import type { ValidationResults } from '../../components/agent_policy_validation';
 
-import { incrementPolicyName } from '../../../services';
+import { incrementPolicyName } from '../../../../services';
 
 import { StepSelectAgentPolicy } from './step_select_agent_policy';
