+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/utils/append_on_save_query_params.test.ts
@@ -6,7 +6,7 @@
  */
 import { parse } from 'query-string';
 
-import type { PackagePolicy, OnSaveQueryParamKeys } from '../../../../types';
+import type { PackagePolicy, OnSaveQueryParamKeys } from '../../../../../types';
 
 import { appendOnSaveQueryParamsToPath } from '.';
