+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/utils/append_on_save_query_params.ts
@@ -11,7 +11,7 @@ import type {
   OnSaveQueryParamOpts,
   PackagePolicy,
   OnSaveQueryParamKeys,
-} from '../../../../types';
+} from '../../../../../types';
 
 export function appendOnSaveQueryParamsToPath({
   path,
