+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/edit_package_policy_page/index.tsx
@@ -49,15 +49,18 @@ import {
 } from '../../../../integrations/hooks';
 import { Loading, Error, ExtensionWrapper, EuiButtonWithTooltip } from '../../../components';
 import { ConfirmDeployAgentPolicyModal } from '../components';
-import { CreatePackagePolicyPageLayout } from '../create_package_policy_page/components';
-import type { PackagePolicyValidationResults } from '../create_package_policy_page/services';
-import { validatePackagePolicy, validationHasErrors } from '../create_package_policy_page/services';
+import { CreatePackagePolicyPageLayout } from '../create_package_policy_page/single_page_layout/components';
+import type { PackagePolicyValidationResults } from '../create_package_policy_page/single_page_layout/services';
+import {
+  validatePackagePolicy,
+  validationHasErrors,
+} from '../create_package_policy_page/single_page_layout/services';
 import type {
   PackagePolicyFormState,
   EditPackagePolicyFrom,
 } from '../create_package_policy_page/types';
-import { StepConfigurePackagePolicy } from '../create_package_policy_page/step_configure_package';
-import { StepDefinePackagePolicy } from '../create_package_policy_page/step_define_package_policy';
+import { StepConfigurePackagePolicy } from '../create_package_policy_page/single_page_layout/step_configure_package';
+import { StepDefinePackagePolicy } from '../create_package_policy_page/single_page_layout/step_define_package_policy';
 import type {
   GetOnePackagePolicyResponse,
   UpgradePackagePolicyDryRunResponse,
